compressed data
raw grid files have gone from f32 to f16, particle coordinates are stored in a single uint32 and particle data in a single uint32

+++ Misc/binaryinfo.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 f = input('Enter file\n')
-a = np.fromfile(f,np.float32,-1)
+a = np.fromfile(f,np.float16,-1)
 a = np.nan_to_num(a,0)
 vals, cou = np.unique(a, return_counts= True)
 print('VALUES')
